requirements.txt
fastapi
uvicorn
requests
torch
langchain
langchain-community
google-generativeai
faiss-cpu
pymupdf
python-docx
python-pptx
beautifulsoup4
pillow
pytesseract
openpyxl
sentence-transformers
pandas